requirements.txt
alembic==1.4.0
aniso8601==8.0.0
APScheduler==3.6.3
autopep8==1.5
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dominate==2.4.0
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Bootstrap==3.3.7.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
httpie==2.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pygtail==0.11.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
tzlocal==2.0.0
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
